Release 24: Iron out the rootless commits
Reduce the amount of resources that any overlaying system can overlay by
removing and hardcoding the main style and main app title string.

As one might notice, the version is now removed from build.gradle and is
solely in AndroidManifest.xml - this makes it easier for us to maintain
updating versions.

Change-Id: Icaa4f929acb32f4c1b1690a18ea319deebad97a0